requirements.txt
sqlalchemy
python-dotenv
pytest
requests
faker